CRUD contributeur

+++ ApiiCollabdev/src/main/java/com/example/ApiiCollabdev/Repository/ContributeurRepository.java
@@ -3,5 +3,5 @@ package com.example.ApiiCollabdev.Repository;
 import com.example.ApiiCollabdev.entities.Contributeur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContributeurRepository extends JpaRepository<Contributeur, Long> {
+public interface ContributeurRepository extends JpaRepository<Contributeur, Integer> {
 }
